Parallelization on a split file.
git-svn-id: svn+ssh://svn.zemanta.com/zdev/neith2/wikiprep/trunk@14910 19fbb8fb-b300-474b-854f-3758a290f079

+++ Wikiprep/Output/Legacy.pm
@@ -7,91 +7,42 @@ use strict;
 
 use File::Path;
 
-use Wikiprep::utils qw( encodeXmlChars getLinkIds removeDuplicatesAndSelf );
+use Wikiprep::utils qw( encodeXmlChars getLinkIds removeDuplicatesAndSelf 
+                        openOutputFile outputFilename openInputFile );
 
-sub new 
-{
+sub new {
 	my $class = shift;
-	my $basepath = shift;
-  my $inputFile = shift;
 
-  my %params = @_;
+  my ($inputFile, %options) = @_;
 
-	my $outputFile = "$basepath.hgw.xml";
+  my $self = {
+               inputFile => $inputFile,
+               part => $options{PART},
+               templateIncDir => &outputFilename($inputFile, ".templates", %options, COMPRESS => 0) 
+             };
 
-  my $anchorTextFile = "$basepath.anchor_text";
+  &_prepareTemplateIncDir(\$self->{templateIncDir});
 
-  # Information about anchor texts for external linnks
-  my $externalAnchorTextFile = "$basepath.external_anchors";
+  if($options{PRESCAN}) {
+    $self->{tempindexf} = IO::File->new("> " . $self->{templateIncDir} . "/index") or die;
+    $self->{tempindexf}->binmode(":utf8");
 
-  my $relatedLinksFile = "$basepath.related_links";
-
-  # Disambiguation links
-  my $disambigPagesFile = "$basepath.disambig";
-
-  # Information about redirects
-  my $redirFile = "$basepath.redir.xml";
-
-  # Information about template inclusion
-  my $templateIncDir = "$basepath.templates";
-
-  my $templateIndexFile = "$templateIncDir/index";
-
-	my $self = {  inputFile             => $inputFile,
-                redirFile             => $redirFile,
-                templateIncDir        => $templateIncDir,
-                statCategoriesFile    => "$basepath.stat.categories",
-                statIncomingLinksFile => "$basepath.stat.inlinks", 
-                catHierarchyFile      => "$basepath.cat_hier" };
-
-  &_prepareTemplateIncDir(\$templateIncDir);
-
-  if( $params{COMPRESS} ) {
-    open( $self->{outf}, "| gzip >$outputFile.gz") or 
-      die "Cannot open pipe to gzip: $!: $outputFile.gz";
-
-    open( $self->{anchorf}, "| gzip > $anchorTextFile.gz") or 
-      die "Cannot open pipe to gzip: $!: $anchorTextFile.gz";
-
-    open( $self->{exanchorf}, "| gzip > $externalAnchorTextFile.gz") or
-      die "Cannot open pipe to gzip: $!: $externalAnchorTextFile.gz";
+    print {$self->{tempindexf}} "# Line format: <Template page id>  <Template name>\n";
   } else {
-    open( $self->{outf}, "> $outputFile") or 
-      die "Cannot open $outputFile: $!";
-    open( $self->{anchorf}, "> $anchorTextFile") or 
-      die "Cannot open $anchorTextFile: $!";
-    open( $self->{exanchorf}, "> $externalAnchorTextFile") or 
-      die "Cannot open $externalAnchorTextFile: $!";
+    $self->{outf}      = openOutputFile($inputFile, ".hgw.xml", %options);
+    $self->{anchorf}   = openOutputFile($inputFile, ".anchor_text", %options);
+    $self->{exanchorf} = openOutputFile($inputFile, ".external_anchors", %options);
+    $self->{relatedf}  = openOutputFile($inputFile, ".related_links", %options, COMPRESS => 0);
+    $self->{disambigf} = openOutputFile($inputFile, ".disambig", %options, COMPRESS => 0);
+
+    &_copyXmlFileHeader($self);
+    print {$self->{anchorf}} "# Line format: <Target page id>  <Source page id>  <Anchor location within text>  <Anchor text (up to the end of the line)>\n\n\n";
+    print {$self->{relatedf}}"# Line format: <Page id>  <List of ids of related articles>\n\n\n";
+    print {$self->{disambigf}} "# Line format: <Disambig page id>  <Target page id (or \"undef\")> <Target anchor> ...\n\n\n";
+    print {$self->{exanchorf}} "# Line format: <Source page id>  <Url>  <Anchor>\n\n\n";
   }
 
-  open( $self->{relatedf}, "> $relatedLinksFile") or 
-    die "Cannot open $relatedLinksFile: $!";
-  
-  open( $self->{disambigf}, "> $disambigPagesFile") or 
-    die "Cannot open $disambigPagesFile: $!";
-
-  open( $self->{tempindexf}, "> $templateIndexFile") or
-    die "Cannot open $templateIndexFile: $!";
-
-  binmode( $self->{outf},       ':utf8');
-  binmode( $self->{anchorf},    ':utf8');
-  binmode( $self->{exanchorf},  ':utf8');
-
-  binmode( $self->{relatedf},   ':utf8');
-  binmode( $self->{disambigf},  ':utf8');
-  binmode( $self->{tempindexf}, ':utf8');
-
-  print {$self->{anchorf}} "# Line format: <Target page id>  <Source page id>  <Anchor location within text>  <Anchor text (up to the end of the line)>\n\n\n";
-  print {$self->{relatedf}}"# Line format: <Page id>  <List of ids of related articles>\n\n\n";
-
-  print {$self->{disambigf}} "# Line format: <Disambig page id>  <Target page id (or \"undef\")> <Target anchor> ...\n\n\n";
-  print {$self->{exanchorf}} "# Line format: <Source page id>  <Url>  <Anchor>\n\n\n";
-  print {$self->{tempindexf}} "# Line format: <Template page id>  <Template name>\n";
-
 	bless $self, $class;
-
-  $self->_copyXmlFileHeader();
-
 	return $self;
 }
 
@@ -99,20 +50,11 @@ sub finish
 {
   my $self = shift;
 
-  print {$self->{outf}} "</mediawiki>\n";
-  close($self->{outf});
+  print {$self->{outf}} "</mediawiki>\n" if $self->{outf};
 
-  close($self->{anchorf});
-  close($self->{exanchorf});
-
-  close($self->{relatedf});
-
-  close($self->{disambigf});
-  close($self->{tempindexf});
-}
-
-sub writeInterwiki
-{
+  while( my ($name, $value) = each(%$self) ) {
+    $value->close if( $name =~ /f$/ );
+  }
 }
 
 # Save information about redirects into an XML-formatted file.
@@ -126,11 +68,10 @@ sub writeRedirects
   my $fromId;
   my $toId;
 
-  open(REDIRF, "> $self->{redirFile}") or die "Cannot open $self->{redirFile}: $!";
-  binmode(REDIRF, ':utf8');
+  my $fh = &openOutputFile($self->{inputFile}, ".redir.xml", PART => $self->{part});
 
-  print REDIRF "<?xml version=\"1.0\" encoding=\"utf-8\"?>\n";
-  print REDIRF "<redirects>\n";
+  print $fh "<?xml version=\"1.0\" encoding=\"utf-8\"?>\n";
+  print $fh "<redirects>\n";
 
   while(($fromTitle, $toTitle) = each(%$refToRedir)) {
 
@@ -153,12 +94,12 @@ sub writeRedirects
     my $encodedToTitle = $toTitle;
     &encodeXmlChars(\$encodedToTitle);
 
-    print REDIRF "<redirect>\n<from>\n<id>", $fromId, "</id>\n<title>", $encodedFromTitle, "</title>\n</from>\n<to>\n<id>", $toId, "</id>\n<title>", $encodedToTitle, "</title>\n</to>\n</redirect>\n"
+    print $fh "<redirect>\n<from>\n<id>", $fromId, "</id>\n<title>", $encodedFromTitle, "</title>\n</from>\n<to>\n<id>", $toId, "</id>\n<title>", $encodedToTitle, "</title>\n</to>\n</redirect>\n"
   }
 
-  print REDIRF "</redirects>\n";
+  print $fh "</redirects>\n";
 	
-  close(REDIRF)
+  close($fh);
 }
 
 sub newTemplate 
@@ -173,15 +114,10 @@ sub _copyXmlFileHeader
 {
   my $self = shift;
 
-  if ($self->{inputFile} =~ /\.gz$/) {
-    open(INF, "gzip -dc $self->{inputFile}|") or die "Cannot open $self->{inputFile}: $!";
-  } elsif ($self->{inputFile} =~ /\.bz2$/) {
-    open(INF, "bzip2 -dc $self->{inputFile}|") or die "Cannot open $self->{inputFile}: $!";
-  } else {
-    open(INF, "< $self->{inputFile}") or die "Cannot open $self->{inputFile}: $!";
-  }
+  my $fh = &openInputFile($self->{inputFile});
 
-  while (<INF>) { # copy lines up to "</siteinfo>"
+  while (<$fh>) { 
+    # copy lines up to "</siteinfo>"
     if (/^<mediawiki /) {
       # The top level element - mediawiki - contains a lot of attributes (e.g., schema)
       # that are no longer applicable to the XML file after our transformation.
@@ -194,7 +130,8 @@ sub _copyXmlFileHeader
     last if (/<\/siteinfo>/);
   }
 
-  close(INF); # this file will later be reopened by "Parse::MediaWikiDump"
+  # this file will later be reopened by "Parse::MediaWikiDump"
+  close($fh);
 }
 
 sub newPage 
@@ -209,65 +146,6 @@ sub newPage
   $self->_logDisambig($page);
 
   $self->_writePage($page);
-}
-
-sub writeStatistics {
-  my $self = shift;
-  my ($refToStatCategories, $refToStatIncomingLinks) = @_;
-
-  my $statCategoriesFile = $self->{statCategoriesFile};
-  my $statIncomingLinksFile = $self->{statIncomingLinksFile};
-
-  open(STAT_CATS, "> $statCategoriesFile") or die "Cannot open $statCategoriesFile";
-  print STAT_CATS "# Line format: <CategoryId (= page id)>  <Number of pages in this category>\n",
-                  "# Here we count the *pages* that belong to this category, i.e., articles AND\n",
-                  "# sub-categories of this category (but not the articles in the sub-categories).\n",
-                  "\n\n";
-
-  my $cat;
-#  foreach $cat ( sort { $statCategories{$b} <=> $statCategories{$a} }
-#                 keys(%statCategories) ) {
-#    print STAT_CATS "$cat\t$statCategories{$cat}\n";
-#  }
-  foreach $cat ( keys(%$refToStatCategories) ) {
-    print STAT_CATS "$cat\t$refToStatCategories->{$cat}\n";
-  }
-  close(STAT_CATS);
-
-  open(STAT_INLINKS, "> $statIncomingLinksFile") or die "Cannot open $statIncomingLinksFile";
-  print STAT_INLINKS "# Line format: <Target page id>  <Number of links to it from other pages>\n\n\n";
-
-  my $destination;
-#  foreach $destination ( sort { $statIncomingLinks{$b} <=> $statIncomingLinks{$a} }
-#                         keys(%statIncomingLinks) ) {
-#    print STAT_INLINKS "$destination\t$statIncomingLinks{$destination}\n";
-#  }
-  foreach $destination ( keys(%$refToStatIncomingLinks) ) {
-    print STAT_INLINKS "$destination\t$refToStatIncomingLinks->{$destination}\n";
-  }
-
-  close(STAT_INLINKS);
-}
-
-sub writeCategoryHierarchy {
-  my $self = shift;
-  my ($refCatHierarchy) = @_;
-
-  my $catHierarchyFile = $self->{catHierarchyFile};
-
-  open(CAT_HIER, "> $catHierarchyFile") or die "Cannot open $catHierarchyFile";
-  print CAT_HIER "# Line format: <Category id>  <List of ids of immediate descendants>\n\n\n";
-
-  my $cat;
-#  foreach $cat ( sort { $catHierarchy{$a} <=> $catHierarchy{$b} }
-#                 keys(%catHierarchy) ) {
-#    print CAT_HIER "$cat\t", join(" ", @{$catHierarchy{$cat}}), "\n";
-#  }
-  foreach $cat ( keys(%$refCatHierarchy) ) {
-    print CAT_HIER "$cat\t", join(" ", @{$refCatHierarchy->{$cat}}), "\n";
-  }
-
-  close(CAT_HIER);
 }
 
 sub _logAnchorText 
@@ -434,7 +312,7 @@ sub _prepareTemplateIncDir(\$) {
   my $n = 1;
   do {
     my $path = "$$refToTemplateIncDir/$n";
-    if( -d $path ) {
+    if( $$refToTemplateIncDir and -d $path ) {
       File::Path::rmtree($path, 0, 0);
     }
     mkdir("$$refToTemplateIncDir/$n");
